Move calculate_bounds_2d back to bevy_sprite
